requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bitarray==3.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
cytoolz==1.0.1
distro==1.9.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
fastapi==0.115.9
frozenlist==1.5.0
google-generativeai==0.4.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
multidict==6.1.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.65.1
parsimonious==0.10.0
propcache==0.3.0
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rlp==4.1.0
scikit-learn==1.4.1.post1
scipy==1.13.1
sentence-transformers==2.7.0
sniffio==1.3.1
starlette==0.45.3
torch==2.3.1
toolz==1.0.0
tqdm==4.67.1
transformers==4.42.4
tweepy==4.15.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
web3==7.8.0
websockets==13.1
yarl==1.18.3
